# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API documentation for the Translation Management Service",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/locales": {
            "get": {
                "tags": [
                    "Locale"
                ],
                "summary": "Get a paginated list of locales",
                "operationId": "99e0538d08dc549df4fd96ee082b1d38",
                "responses": {
                    "200": {
                        "description": "Language retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Locale"
                ],
                "summary": "Store a new locale",
                "operationId": "0fae2a12a32d1c38e6f8cb865e68a82b",
                "requestBody": {
                    "description": "Pass locale data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "English"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Language created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/locales/{locale}": {
            "get": {
                "tags": [
                    "Locale"
                ],
                "summary": "Get a single locale",
                "operationId": "2e7f3c774bbcdf0504a16fb67540999f",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "ID of locale to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Language retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Locale"
                ],
                "summary": "Update an existing locale",
                "operationId": "b9f47f4c43e756285d77a8df546d86e4",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "ID of locale to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass updated locale data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "fr"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "French"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Language updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locale"
                ],
                "summary": "Delete a locale",
                "operationId": "fff4f32edffa47812bb765e70a0f5b8f",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "ID of locale to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Language deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}